fix error handling and collapsible trigger
